🔧 Refactor: Update import paths to new service module structure and add function_services implementation
