+++ services/catalog/src/simcore_service_catalog/api/rest/_services_specifications.py
@@ -12,7 +12,7 @@ from models_library.users import UserID
 from ..._constants import RESPONSE_MODEL_POLICY
 from ...repository.groups import GroupsRepository
 from ...repository.services import ServicesRepository
-from ...services.function_services import is_function_service
+from ...service.function_services import is_function_service
 from ..dependencies.database import get_repository
 from ..dependencies.services import get_default_service_specifications
