+++ services/catalog/src/simcore_service_catalog/api/rpc/_services.py
@@ -26,7 +26,7 @@ from servicelib.rabbitmq.rpc_interfaces.catalog.errors import (
 from simcore_service_catalog.repository.groups import GroupsRepository
 
 from ...repository.services import ServicesRepository
-from ...services import services_api
+from ...service import services_api
 from ..dependencies.director import get_director_api
 
 _logger = logging.getLogger(__name__)
