+++ services/catalog/src/simcore_service_catalog/service/compatibility.py
