+++ services/catalog/src/simcore_service_catalog/service/manifest.py
